Move both Framing filters into a single file
