--- linkerd/protocol/http/src/main/scala/io/buoyant/linkerd/protocol/FramingException.scala
@@ -1,6 +1,0 @@
-package io.buoyant.linkerd.protocol
-
-import io.buoyant.linkerd.ProtocolException
-
-case class FramingException(reason: String)
-  extends ProtocolException(reason, "HTTP")
