+++ linkerd/protocol/http/src/main/scala/io/buoyant/linkerd/protocol/HttpConfig.scala
@@ -34,7 +34,7 @@ class HttpInitializer extends ProtocolInitializer.Simple {
       .prepend(Headers.Dst.PathFilter.module)
       .replace(StackClient.Role.prepFactory, DelayedRelease.module)
       .prepend(http.ErrorResponder.module)
-      .insertAfter(http.ErrorResponder.role, RequestFramingFilter.module)
+      .insertAfter(http.ErrorResponder.role, FramingFilter.ServerFilter.module)
     val boundStack = Http.router.boundStack
       .prepend(Headers.Dst.BoundFilter.module)
     val clientStack = Http.router.clientStack
@@ -44,7 +44,7 @@ class HttpInitializer extends ProtocolInitializer.Simple {
       .insertAfter(DtabStatsFilter.role, HttpLoggerConfig.module)
       .insertAfter(Retries.Role, http.StatusCodeStatsFilter.module)
       .insertAfter(AddrMetadataExtraction.Role, RewriteHostHeader.module)
-      .insertAfter(FailureAccrualFactory.role, ResponseFramingFilter.module)
+      .insertAfter(FailureAccrualFactory.role, FramingFilter.ClientFilter.module)
 
     Http.router
       .withPathStack(pathStack)
